+++ examples/resources/iosxr_vrf_route_target_four_byte_as_format/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_vrf_route_target_four_byte_as_format.example "Cisco-IOS-XR-um-vrf-cfg:/vrfs/vrf[vrf-name=VRF1]/address-family/ipv4/unicast/Cisco-IOS-XR-um-router-bgp-cfg:import/Cisco-IOS-XR-um-router-bgp-cfg:route-target/Cisco-IOS-XR-um-router-bgp-cfg:four-byte-as-rts/Cisco-IOS-XR-um-router-bgp-cfg:four-byte-as-rt[as-number=100000][index=1][stitching=true]"