+++ docs/guides/changelog.md
@@ -13,6 +13,10 @@ description: |-
 - Allow provider config without host attribute in case devices attribute is being used
 - Enhance iosxr_gnmi resource to support nested attributes (within YANG containers) using "/" as separator
 - Enhance iosxr_gnmi resource to support nested YANG lists
+- BREAKING CHANGE: remove iosxr_vrf_route_target_two_byte_as_format resource and data source
+- BREAKING CHANGE: remove iosxr_vrf_route_target_four_byte_as_format resource and data source
+- BREAKING CHANGE: remove iosxr_vrf_route_target_ip_address_format resource and data source
+- Add route target attributes to iosxr_vrf resource and data source
 
 ## 0.1.0